requirements.txt
slack-sdk==3.35.0
PyYAML==6.0.2
